backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
beautifulsoup4==4.12.3
langdetect==1.0.9
Pillow==10.2.0
requests==2.31.0
